rename styled file
